+++ packages/orq-rc/RELEASES.md
@@ -2919,3 +2919,13 @@ Based on:
 - [typescript v3.11.0-rc.2] packages/orq-rc
 ### Releases
 - [NPM v3.11.0-rc.2] https://www.npmjs.com/package/@orq-ai/node/v/3.11.0-rc.2 - packages/orq-rc
+
+## 2025-08-08 08:04:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.599.0 (2.675.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.11.0-rc.3] packages/orq-rc
+### Releases
+- [NPM v3.11.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/3.11.0-rc.3 - packages/orq-rc